Crate gd_props

Methods and functions to to save and load your GodotClass Resources using serde.

Resources defined in Rust using [godot] crate are fully dependent on default Godot load/save rules - saved to .tres files, which requires all saveable properties to be Godot-recognized types annotated with #[export] attribute. gd-props provides framework to make their saving and loading independent of Godot rules, using serde ruleset instead.

It provides two new custom formats, that the Resources will be saved to:

.gdron - human readable format based on Ron serialization format from ron crate.

.gdbin - binary format based on MessagePack serialization fromat from [rmp_serde] crate.

The core functionality is based in three derive macros:

GdProp - used to implement GdProp trait to the user-defined Resource, making it saveable and loadable to/from .gdron and .gdbin files.

GdPropLoader and GdPropSaver - used to implement GdPropLoader and GdPropSaver traits to user-defined ResourceFormatLoader and ResourceFormatSaver, which will be used by Godot to load and save GdProp-annotated resources.

Additionally, crate::serde_gd module contains submodules to be used with #[serde(with)] attribute macro, making it possible to serialize sub-resources contained within GdProp-annotated resource.

Modules

errors

serde_gd

Module containing serialization and deserialization methods for Godot objects and their collections.

traits

Module containing traits implemented by provided macros. There shouldn’t be a necessity to implement them directly by the user.

types

Supplementary types for easier handling of pointers of Resource-inheriting GodotClasses.

Derive Macros

GdProp

Macro used to implement GdProp trait, which makes a rust-defined Godot Resource saveable and loadable to and from .gdron and .gdbin formats.

GdPropLoader

Create resource loader for GdProp resources.

GdPropSaver

Create resource saver for GdProp resources.