Attribute Macro gd_props::gd_props_plugin

#[gd_props_plugin]

Implement GodotClasses necessary for .gdbin and .gdron files handling within Godot

This single macro is used to implement four different, intertwined GodotClass structs with all necessary implementations:

ResourceFormatLoader and ResourceFormatSaver: used to load and save GdProp-deriving resources to .gdbin and .gdron formats,

EditorPlugin and EditorExportPlugin which handle exporting .gdbin and .gdron format files. .gdron files are transformed into more compact and faster .gdbin format during export.

Identifiers will be generated based on provided struct Identifier, with the visibility marker provided, either pub or pub(crate):

EditorPlugin: Identifier,

EditorExportPlugin: IdentifierExporter,

ResourceFormatSaver: IdentifierSaver,

ResourceFormatLoader: IdentifierLoader.

Register GdProp resources

Every resource that should be saveable/loadable/exportable as .gdbin/.gdron file needs to be provided in helper #[register] macro attribute, as seen in example below. Multiple #[register] helper macros with different identifiers can be provided for code readability.

Setup

Created plugins don’t need further setup: as they are created, they will be registered and used by Godot automatically during export.

Loader and Saver need registering in your #[gdextension] implementation. It is recommended to use provided associated functions: register_saver and register_loader - for implementation details see their documentation.

Example

use godot::prelude::*;
use gd_props::gd_props_plugin;

// Macro creates four different GodotClasses and registers two resources implementing `GdProp`
#[gd_props_plugin]
#[register(MyResource, MyOtherResource)]
pub(crate) struct MyPropPlugin;

// Plugin and Exporter are only available in-editor for exporting resources.
assert_eq!(MyPropPlugin::INIT_LEVEL, InitLevel::Editor);
assert_eq!(MyPropPluginExporter::INIT_LEVEL, InitLevel::Editor);

// Loader and Saver are available in scenes for loading/saving registered resources.
assert_eq!(MyPropPluginSaver::INIT_LEVEL, InitLevel::Scene);
assert_eq!(MyPropPluginLoader::INIT_LEVEL, InitLevel::Scene);