Module gd_props::serde_gd::gd_array

Module that can be used to serialize and deserialize objects castable to Resource on basis of their pointers contained within Array collection.

Its main use is to derive serde::Serialize and serde::Deserialize on resources containing pointers to other resources, while keeping data from every one in serialized struct.

Example

use godot::prelude::*;
use godot::classes::{Resource, IResource};
use godot::builtin::Array;
use serde::{Serialize, Deserialize};

#[derive(GodotClass, Serialize, Deserialize)]
#[class(base=Resource, init)]
struct InnerResource {}

#[godot_api]
impl InnerResource {}

#[derive(GodotClass, Serialize, Deserialize)]
#[class(init, base=Resource)]
struct OuterResource {
    #[export]
    #[serde(with="gd_props::serde_gd::gd_array")]
    inner: Array<Gd<InnerResource>>
}

#[godot_api]
impl OuterResource {}

Functions

deserialize

serialize