Trait gd_props::traits::GdProp

pub trait GdProp: Serialize + for<'de> Deserialize<'de> + GodotClass + UserClass + Inherits<Resource> {
    const HEAD_IDENT: &'static str;

    // Provided methods
    fn save_bin(&self, path: GString) -> Error { ... }
    fn load_bin(path: GString) -> Variant { ... }
    fn save_ron(&self, path: GString) -> Error { ... }
    fn translate_ron_to_bin(path: GString) -> PackedByteArray { ... }
    fn load_ron(path: GString) -> Variant { ... }
}

GdProp saveable resource

Trait which provides methods to serialize and deserialize rust-defined Resource to:

.gdbin files, based on MessagePack

.gdron files, based on ron

Required Associated Constants

const HEAD_IDENT: &'static str

Struct identifier included in gdron file.

Provided Methods

fn save_bin(&self, path: GString) -> Error

Save object to a file located at path in .gdbin format.

fn load_bin(path: GString) -> Variant

Load object from a file located at path in .gdbin format.

fn save_ron(&self, path: GString) -> Error

Save object to a file located at path in ron format.

fn translate_ron_to_bin(path: GString) -> PackedByteArray

fn load_ron(path: GString) -> Variant

Load object from a file located at path in ron format.

Object Safety

This trait is not object safe.