Crate gd_props

Methods and functions to to save and load your GodotClass Resources using serde.

Resources defined in Rust using [godot] crate are fully dependent on default Godot load/save rules - saved to .tres files, which requires all saveable properties to be Godot-recognized types annotated with #[export] attribute. gd-props provides framework to make their saving and loading independent of Godot rules, using serde ruleset instead.

It provides two new custom formats, that the Resources will be saved to:

.gdron - human readable format based on Ron serialization format from ron crate.

.gdbin - binary format based on MessagePack serialization fromat from [rmp_serde] crate.

The core functionality is based in three derive macros:

GdProp - used to implement GdProp trait to the user-defined Resource, making it saveable and loadable to/from .gdron and .gdbin files.

gd_props_plugin - used to create all needed GodotClass structs to handle GdProp-annotated resources during saving/loading to and from custom formats and during project export.

Additionally, crate::serde_gd module contains submodules to be used with #[serde(with)] attribute macro, making it possible to serialize sub-resources contained within GdProp-annotated resource.

Modules

errors

private

Module containing types necessary for export plugin. Symbols not needed outside of internal usage.

serde_gd

Module containing serialization and deserialization modules for pointers to Godot Resource and their collections.

traits

Module containing traits implemented by provided macros. There shouldn’t be a necessity to implement them directly by the user.

Attribute Macros

gd_props_plugin

Implement GodotClasses necessary for .gdbin and .gdron files handling within Godot

Derive Macros

GdProp

Macro used to implement GdProp trait, which makes a rust-defined Godot Resource saveable and loadable to and from .gdron and .gdbin formats.